Game prefabs
Due to the current simplicity of the project all enemy and player prefabs work using basic scripts for movement, custom made colliders, and life script that provides a health amount that upon reaching zero destroys the prefab. Additionally, the player also counts with a rigidbody component to implement functions such as jump.